Updated Lattice and Added Static Tests
Added new version of generateLattice all other prior to 5_30 should be
considered legacy code and moved to old code folder or removed.
staticFlatAlphaSweep script should be eventual be moved to a testing
folder but was updated for testing, still in main folder.
staticAlphaSweep should be primary static test function, still needs to
be commented